Rename thermoChem.{hpp,cpp} to calorically_perfect.{hpp,cpp}
This better describes what these files contain.
