+++ src/tps.cpp
@@ -56,7 +56,6 @@
 #include "cycle_avg_joule_coupling.hpp"
 #include "independent_coupling.hpp"
 #include "loMach.hpp"
-#include "thermoChem.hpp"
 
 namespace TPS {
